requirements.txt
streamlit
opencv-python-headless
easyocr
numpy
google-generativeai
python-dotenv
Pillow